Move types to separate file; Adjust interfaces
