+++ src/util/types.ts
@@ -1,0 +1,33 @@
+export type CabinType =
+  | 'XL'
+  
+export type ChassisType =
+  | 'XL'
+  
+export type Color =
+  | 'Sunny Yellow'
+
+export type InteriorType =
+  | 'XL'
+
+export type Manufacturer =
+  | 'DAF'
+  | 'Iveco'
+  | 'MAN'
+  | 'Mercedes-Benz'
+  | 'Renault'
+  | 'Scania'
+  | 'Volvo';
+
+export type Model =
+  | 'TGX'
+  | 'FH'
+  | 'Stralis'
+  | 'New Actros'
+  | 'XF'
+  | 'FX'
+  | 'T'
+
+export type WheelType =
+  | '385/55 Michelin X Line Energy F'
+  | '315/70 Michelin X Line Energy D 2'
